Add reference to TrainingPeaks blog post with TSS formula
TrainingPeaks' blog has a post by Joe Friel that has the TSS formula,
and it uses normalized power. Add this to the references as my only
other reference for the TSS formula was the 2003 Coggan post, which
says to use average power, even though I knew normalized power is the
correct value to use.